requirements.txt
pytest
selenium
requests
allure-pytest
config
flake8
